Main.slx - tried increasing sampling frequecy
Test for increasing sampling frequency was done resulting in minor
changes for some simulink-blocks. The change was due to fixed
sampling-time to 0.02. Instead the blocks now take time_step as
sampling-time, making the model more fleximble.

However, the test was unsuccessfull. The killswitch on the RC-controller
stopped working and many poorly sampeled samples was produced. Further
investigation to this should be taken if increased sampling frequency is
needed.  Perhaps some variable in the c-code somewhere?

Sampling-time changed back to 0.02. It works as normal on Hugin